fix invite user
