+++ src/routes/admin/users/+page.server.ts
@@ -1,8 +1,11 @@
 import { Role } from "$lib/schema";
 import { getConfig } from "$lib/server/config";
 import { client } from "$lib/server/prisma";
-import { redirect, error } from "@sveltejs/kit";
-import type { PageServerLoad } from "./$types";
+import { redirect, error, fail } from "@sveltejs/kit";
+import type { Actions, PageServerLoad } from "./$types";
+import { sendSignupLink } from "$lib/server/email";
+import generateToken, { hashToken } from "$lib/server/token";
+import { z } from "zod";
 
 export const load: PageServerLoad = async ({ locals }) => {
 	const { session, user } = await locals.validateUser();
@@ -49,4 +52,51 @@ export const load: PageServerLoad = async ({ locals }) => {
 		})),
 		config
 	};
+};
+
+export const actions: Actions = {
+	default: async ({ url, request }) => {
+		const token = await generateToken();
+		const tokenUrl = new URL(`/signup?token=${token}`, url);
+
+		const config = await getConfig();
+
+		if (!config.smtp.enable) {
+			await client.signupToken.create({
+				data: {
+					hashedToken: hashToken(token),
+					expiresIn: 21600000 // 6 hours in milliseconds
+				}
+			});
+
+			return { action: "invite-email", success: true, url: tokenUrl.href };
+		}
+
+		const formData = Object.fromEntries(await request.formData());
+		const schema = z.object({
+			"invite-email": z.string().email()
+		});
+
+		const emailData = schema.safeParse(formData);
+
+		if (!emailData.success) {
+			const errors = emailData.error.errors.map((error) => {
+				return {
+					field: error.path[0],
+					message: error.message
+				};
+			});
+			return fail(400, { action: "invite-email", error: true, errors });
+		}
+
+		await client.signupToken.create({
+			data: {
+				hashedToken: hashToken(token),
+				expiresIn: 21600000 // 6 hours in milliseconds
+			}
+		});
+
+		await sendSignupLink(emailData.data["invite-email"], tokenUrl.href);
+		return { action: "invite-email", success: true, url: null };
+	}
 };
